tool_orchestration_env/server/requirements.txt
openenv-core[core]>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
pydantic>=2.0.0
